Fetch realtime internal transactions and balances synchronously
Fixes #366

Queuing the internal transactions and balances to be fetched
asynchronously for realtime blocks lead to upto 15 minute delay between
when a block appeared and the internal transaction was drained from the
queue.  It is a better user experience to not see the realtime blocks at
all than to see them incomplete, so fetch the internal transactions and
then the balances before importing the whole tree of data instead of
using the catchup asynchronous queues.