Retry retried_entries for uncles
`Indexer.Block.Uncle.Fetcher.retry/1` when there are errors did not
actually retry the calculated `retried_errors` and instead only logged
them.  This didn't cause an unused variable warning because
`retried_errors` is used in the Logger metadata.  The `Task` for the
`run/2` callback did not error because `Logger.error` returns `:ok`.

To fix this problem, `{:retry, retried_errors}` need to be returned
after the `Logger.error` call.