Optimize token holder count updates when importing address current balances
Problem: realtime fetcher transactions takes too long and often fails with
a timeout. The reason of timeout is deleted_address_current_token_balances
step of import, which selects previous balances for the rows we are going
to update. The SELECT query it makes is extremely slow.

Solution: we incorporate fetching old balance value into the upsert itself.
This is done by help of a transient old_value field, which is updated in
ON CONFLICT clause of INSERT with the previous value and returned.

Upgrading: migration 20190321185644_add_old_value_for_current_token_balances.exs