set correct last calue for coin balances chart data
Sometimes the last value for coin history chart data is
incorrect because when selecting records they are grouped
by day and max value is selected in that day. So if there
are a couple of coin balance changes only max value is shown.

This PR always shows the latest value from the database by
block number.