do not fail block `internal_transactions_indexed_at` field update
this runner updates internal_transactions_index_at field in blocks
table. but before this, it checks if these blocks are consensus, some
blocks may lose consensus.

we just shouldn't check if the number of updated records is exactly
equal to the number of records that we want to insert.
blocks that lost consensus will be re-fetched